Fixed a bug that resulted in printing erroneous version numbers.
It seems that calls to 'glGetIntegerv', which are made when finding
supported OpenGL version numbers, can fail, which then results in
the major and minor version numbers keeping their initial values.
I've now set these variables to initialize to zero to prevent wild
and meaningless values from appearing in the console when trying to
load the library of functions.